+++ module/rust/proc_macro_tools/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "proc_macro_tools"
-version = "0.1.7"
+version = "0.1.8"
 edition = "2021"
 authors = [
   "Kostiantyn Wandalen <wandalen@obox.systems>",
