+++ module/rust/type_constructor/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "type_constructor"
-version = "0.1.18"
+version = "0.1.19"
 edition = "2021"
 authors = [
   "Kostiantyn Wandalen <wandalen@obox.systems>",
