+++ rust/test/dt/type_constructor/dynamic/make/make_too_many.rs
@@ -5,5 +5,3 @@ fn main()
 {
   let x = make!( 0, 1, 2, 3 );
 }
-
-/* xxx : does not work properly currently */
