+++ rust/test/dt/type_constructor/inc.rs
@@ -2,7 +2,14 @@
 #[ allow( unused_imports ) ]
 use super::*;
 
-mod single_test;
+mod single_parameter_main_gen_test;
+mod single_parameter_main_manual_test;
+mod single_parameter_test;
+mod single_parametrized_main_gen_test;
+mod single_parametrized_main_manual_test;
+mod single_parametrized_test;
+
+/* qqq : for Dima : split this tests */
 mod pair_test;
 mod homo_pair_test;
 
@@ -33,4 +40,3 @@ mod make_interface_test;
 mod vectorized_from_test;
 
 mod enumerable_test;
-
