+++ rust/test/dt/type_constructor/many_parameter_main_test_only.rs
@@ -1,3 +1,17 @@
+#[ derive( PartialEq, Debug ) ]
+struct MySingle
+(
+  pub f32,
+);
+impl From< MySingle >
+for f32
+{
+  fn from( src : MySingle ) -> Self
+  {
+    src.0
+  }
+}
+
 tests_impls!
 {
 
@@ -21,7 +35,7 @@ tests_impls!
     }
 
     /* test.case( "basic" ) */
-    let instance1 = Many::< f32 >::from( 13.0 );
+    let instance1 = Many::< f32 >::from( core::iter::once( 13.0_f32 ) );
     assert!( implements!( instance1 => PartialEq ) );
     assert!( implements!( instance1 => Clone ) );
     assert!( implements!( instance1 => fmt::Debug ) );
@@ -52,64 +66,82 @@ tests_impls!
     }
 
     /* test.case( "from f32 into Many" ) */
-    let instance1 : Many< f32 > = ( 13.0 ).into();
-    let instance2 = Many::< f32 >::from( 13.0 );
+    let instance1 : Many< f32 > = ( core::iter::once( 13.0 ) ).into();
+    let instance2 = Many::< f32 >::from( core::iter::once( 13.0 ) );
     a_id!( instance1.0, vec!( 13.0 ) );
     a_id!( instance2.0, vec!( 13.0 ) );
     a_id!( instance1, instance2 );
 
-    /* test.case( "from &f32 into Many" ) */
-    let instance1 : Many< f32 > = ( &13.0 ).into();
-    let instance2 = Many::< f32 >::from( &13.0 );
-    a_id!( instance1.0, vec!( 13.0 ) );
-    a_id!( instance2.0, vec!( 13.0 ) );
-    a_id!( instance1, instance2 );
+    // /* test.case( "from &f32 into Many" ) */
+    // let instance1 : Many< f32 > = ( &13.0 ).into();
+    // let instance2 = Many::< f32 >::from( &13.0 );
+    // a_id!( instance1.0, vec!( 13.0 ) );
+    // a_id!( instance2.0, vec!( 13.0 ) );
+    // a_id!( instance1, instance2 );
 
     /* test.case( "from itself into itself" ) */
-    let instance1 : Many< f32 > = ( Many::from( 13.0 ) ).into();
-    let instance2 = Many::< f32 >::from( Many::from( 13.0 ) );
+    let instance1 : Many< f32 > = ( Many::from( core::iter::once( 13.0 ) ) ).into();
+    let instance2 = Many::< f32 >::from( Many::from( core::iter::once( 13.0 ) ) );
     a_id!( instance1.0, vec!( 13.0 ) );
     a_id!( instance2.0, vec!( 13.0 ) );
     a_id!( instance1, instance2 );
 
-    /* test.case( "from tuple" ) */
-    let got : Many< f32 > = ( 13.0, ).into();
-    a_id!( got, Many::from( 13.0 ) );
-    let got = Many::< f32 >::from( ( 13.0, ) );
-    a_id!( got, Many::from( 13.0 ) );
+    // /* test.case( "from tuple" ) */
+    // let got : Many< f32 > = ( 13.0, ).into();
+    // a_id!( got, Many::from( core::iter::once( 13.0 ) ) );
+    // let got = Many::< f32 >::from( ( 13.0, ) );
+    // a_id!( got, Many::from( core::iter::once( 13.0 ) ) );
 
     /* test.case( "from array" ) */
-    let got : Many< f32 > = [ 13.0 ].into();
-    a_id!( got, Many::from( 13.0 ) );
-    let got = Many::< f32 >::from( [ 13.0 ] );
-    a_id!( got, Many::from( 13.0 ) );
+    let got : Many< f32 > = [ 1.0, 3.0 ].into();
+    a_id!( got, Many( vec![ 1.0, 3.0 ] ) );
+    let got = Many::< f32 >::from( [ 1.0, 3.0 ] );
+    a_id!( got, Many( vec![ 1.0, 3.0 ] ) );
+
+    /* test.case( "from array of singles" ) */
+    let got : Many< f32 > = [ MySingle( 1.0 ), MySingle( 3.0 ) ].into();
+    a_id!( got, Many( vec![ 1.0, 3.0 ] ) );
+    let got = Many::< f32 >::from( [ MySingle( 1.0 ), MySingle( 3.0 ) ] );
+    a_id!( got, Many( vec![ 1.0, 3.0 ] ) );
+
+    /* test.case( "from list" ) */
+    let got : Many< f32 > = vec![ 1.0, 3.0 ].into();
+    a_id!( got, Many( vec![ 1.0, 3.0 ] ) );
+    let got = Many::< f32 >::from( vec![ 1.0, 3.0 ] );
+    a_id!( got, Many( vec![ 1.0, 3.0 ] ) );
+
+    /* test.case( "from list of singles" ) */
+    let got : Many< f32 > = vec![ MySingle( 1.0 ), MySingle( 3.0 ) ].into();
+    a_id!( got, Many( vec![ 1.0, 3.0 ] ) );
+    let got = Many::< f32 >::from( vec![ MySingle( 1.0 ), MySingle( 3.0 ) ] );
+    a_id!( got, Many( vec![ 1.0, 3.0 ] ) );
 
     /* test.case( "from slice" ) */
-    let got : Many< f32 > = (&[ 13.0 ][ .. ]).into();
-    a_id!( got, Many::from( 13.0 ) );
-    let got = Many::< f32 >::from( (&[ 13.0 ][ .. ]) );
-    a_id!( got, Many::from( 13.0 ) );
+    let got : Many< f32 > = (&[ 13.0 ][ .. ]).iter().cloned().into();
+    a_id!( got, Many::from( core::iter::once( 13.0 ) ) );
+    let got = Many::< f32 >::from( (&[ 13.0 ][ .. ]).iter().cloned() );
+    a_id!( got, Many::from( core::iter::once( 13.0 ) ) );
 
     /* test.case( "from slice" ) */
-    let got : Many< f32 > = (&[ 1.0, 2.0, 3.0 ][ .. ]).into();
+    let got : Many< f32 > = (&[ 1.0, 2.0, 3.0 ][ .. ]).iter().cloned().into();
     a_id!( got, Many::from( [ 1.0, 2.0, 3.0 ] ) );
-    let got = Many::< f32 >::from( (&[ 1.0, 2.0, 3.0 ][ .. ]) );
+    let got = Many::< f32 >::from( (&[ 1.0, 2.0, 3.0 ][ .. ]).iter().cloned() );
     a_id!( got, Many::from( [ 1.0, 2.0, 3.0 ] ) );
 
     /* test.case( "clone / eq" ) */
-    let instance1 : Many< f32 > = ( 13.0 ).into();
+    let instance1 : Many< f32 > = core::iter::once( 13.0 ).into();
     let instance2 = instance1.clone();
     a_id!( instance2.0, vec!( 13.0 ) );
     a_id!( instance1, instance2 );
 
     /* test.case( "deref" ) */
-    let mut got : Many< f32 > = ( 13.0 ).into();
+    let mut got : Many< f32 > = core::iter::once( 13.0 ).into();
     a_id!( got.len(), 1 );
     a_id!( got.pop(), Some( 13.0 ) );
     a_id!( got.0, std::vec::Vec::< f32 >::new() );
 
     /* test.case( "as_slice" ) */
-    let src : Many< f32 > = ( 13.0, ).into();
+    let src : Many< f32 > = core::iter::once( 13.0 ).into();
     let got = src.as_slice();
     a_id!( got, &[ 13.0, ][ .. ] );
     assert!( !mem::same_ptr( &src, got ) );
