+++ rust/test/dt/type_constructor/many_parametrized_main_test_only.rs
@@ -1,3 +1,17 @@
+// #[ derive( PartialEq, Debug ) ]
+// struct MySingle
+// (
+//   pub f32,
+// );
+// impl From< MySingle >
+// for f32
+// {
+//   fn from( src : MySingle ) -> Self
+//   {
+//     src.0
+//   }
+// }
+
 tests_impls!
 {
 
@@ -37,12 +51,12 @@ tests_impls!
     a_id!( instance2.0, vec![ mk!( 13.0 ) ] );
     a_id!( instance1, instance2 );
 
-    /* test.case( "from &f32 into Many" ) */
-    let instance1 : Many< f32, f64 > = ( &mk!( 13.0 ) ).into();
-    let instance2 = Many::< f32, f64 >::from( &mk!( 13.0 ) );
-    a_id!( instance1.0, vec![ mk!( 13.0 ) ] );
-    a_id!( instance2.0, vec![ mk!( 13.0 ) ] );
-    a_id!( instance1, instance2 );
+    // /* test.case( "from &f32 into Many" ) */
+    // let instance1 : Many< f32, f64 > = ( &mk!( 13.0 ) ).into();
+    // let instance2 = Many::< f32, f64 >::from( &mk!( 13.0 ) );
+    // a_id!( instance1.0, vec![ mk!( 13.0 ) ] );
+    // a_id!( instance2.0, vec![ mk!( 13.0 ) ] );
+    // a_id!( instance1, instance2 );
     // yyy
 
     /* test.case( "from itself into itself" ) */
@@ -52,20 +66,14 @@ tests_impls!
     a_id!( instance2.0, vec![ mk!( 13.0 ) ] );
     a_id!( instance1, instance2 );
 
-    // /* test.case( "from vector" ) */
-    // let got : Many< f32, f64 > = vec![ mk!( 13.0 ), mk!( 31.0 ) ].into();
-    // let exp : Many< f32, f64 > = Many::from( vec![ mk!( 13.0 ), mk!( 31.0 ) ] );
+    // /* test.case( "from tuple" ) */
+    // let got : Many< f32, f64 > = ( mk!( 13.0 ), ).into();
+    // let exp : Many< f32, f64 > = Many::from([ mk!( 13.0 ) ]);
     // a_id!( got, exp );
-    // xxx
-
-    /* test.case( "from tuple" ) */
-    let got : Many< f32, f64 > = ( mk!( 13.0 ), ).into();
-    let exp : Many< f32, f64 > = Many::from([ mk!( 13.0 ) ]);
-    a_id!( got, exp );
-    let got = Many::< f32, f64 >::from( ( mk!( 13.0 ), ) );
-    let exp : Many< f32, f64 > = Many::from([ mk!( 13.0 ) ]);
-    a_id!( got, exp );
-// yyy
+    // let got = Many::< f32, f64 >::from( ( mk!( 13.0 ), ) );
+    // let exp : Many< f32, f64 > = Many::from([ mk!( 13.0 ) ]);
+    // a_id!( got, exp );
+    // yyy
 
     /* test.case( "from array" ) */
     let got : Many< f32, f64 > = [ mk!( 13.0 ), ].into();
@@ -83,27 +91,37 @@ tests_impls!
     let exp : Many< f32, f64 > = Many::from( [ mk!( 1.0 ), mk!( 2.0 ), mk!( 3.0 ) ] );
     a_id!( got, exp );
 
-    // // xxx
-    // let slice = &[ mk!( 13.0 ), ][ .. ];
-    // for e in slice
-    // {
-    //   inspect_type::inspect_type_of!( e )
-    //   // dbg!( e );
-    // }
+    /* test.case( "from array of singles" ) */
+    let got : Many< f32, f64 > = [ 1.0, 3.0 ].into();
+    a_id!( got, Many( vec![ mk!( 1.0 ), mk!( 3.0 ) ] ) );
+    let got = Many::< f32, f64 >::from( [ 1.0, 3.0 ] );
+    a_id!( got, Many( vec![ mk!( 1.0 ), mk!( 3.0 ) ] ) );
+
+    /* test.case( "from list" ) */
+    let got : Many< f32, f64 > = vec![ mk!( 1.0 ), mk!( 3.0 ) ].into();
+    a_id!( got, Many( vec![ mk!( 1.0 ), mk!( 3.0 ) ] ) );
+    let got = Many::< f32, f64 >::from( vec![ mk!( 1.0 ), mk!( 3.0 ) ] );
+    a_id!( got, Many( vec![ mk!( 1.0 ), mk!( 3.0 ) ] ) );
+
+    /* test.case( "from list of singles" ) */
+    let got : Many< f32, f64 > = vec![ 1.0, 3.0 ].into();
+    a_id!( got, Many( vec![ mk!( 1.0 ), mk!( 3.0 ) ] ) );
+    let got = Many::< f32, f64 >::from( vec![ 1.0, 3.0 ] );
+    a_id!( got, Many( vec![ mk!( 1.0 ), mk!( 3.0 ) ] ) );
 
     /* test.case( "from slice" ) */
-    let got : Many< f32, f64 > = ( &[ mk!( 13.0 ), ][ .. ] ).into();
+    let got : Many< f32, f64 > = ( ( &[ mk!( 13.0 ), ][ .. ] ).iter().cloned() ).into();
     let exp : Many< f32, f64 > = Many::from([ mk!( 13.0 ) ]);
     a_id!( got, exp );
-    let got = Many::< f32, f64 >::from( &[ mk!( 13.0 ), ][ .. ] );
+    let got = Many::< f32, f64 >::from( ( &[ mk!( 13.0 ), ][ .. ] ).iter().cloned() );
     let exp : Many< f32, f64 > = Many::from([ mk!( 13.0 ) ]);
     a_id!( got, exp );
 
     /* test.case( "from slice" ) */
-    let got : Many< f32, f64 > = ( &[ mk!( 1.0 ), mk!( 2.0 ), mk!( 3.0 ) ][ .. ] ).into();
+    let got : Many< f32, f64 > = ( &[ mk!( 1.0 ), mk!( 2.0 ), mk!( 3.0 ) ][ .. ] ).iter().cloned().into();
     let exp : Many< f32, f64 > = Many::from( [ mk!( 1.0 ), mk!( 2.0 ), mk!( 3.0 ) ] );
     a_id!( got, exp );
-    let got = Many::< f32, f64 >::from( &[ mk!( 1.0 ), mk!( 2.0 ), mk!( 3.0 ) ][ .. ] );
+    let got = Many::< f32, f64 >::from( ( &[ mk!( 1.0 ), mk!( 2.0 ), mk!( 3.0 ) ][ .. ] ).iter().cloned() );
     let exp : Many< f32, f64 > = Many::from( [ mk!( 1.0 ), mk!( 2.0 ), mk!( 3.0 ) ] );
     a_id!( got, exp );
