+++ rust/test/dt/type_constructor/single_parameter_main_gen_test.rs
@@ -1,0 +1,11 @@
+#[ allow( unused_imports ) ]
+use super::*;
+
+TheModule::types!
+{
+  #[ derive( Debug, Clone ) ]
+  #[ derive( PartialEq, Default ) ]
+  single Single : < T >;
+}
+
+include!( "./single_parameter_main_test_only.rs" );
